refactor: split constants file into multiple
